requirements.txt
certifi==2020.6.20
joblib==0.16.0
llvmlite==0.34.0
numba==0.51.2
numpy==1.19.2
pandas==1.1.2
pynndescent==0.4.8
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
threadpoolctl==2.1.0
umap-learn==0.4.6
